Adds disable-upgrade functionality directly to LegacyElementMixin
