+++ lib/legacy/class.js
@@ -9,11 +9,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 */
 
 import { LegacyElementMixin } from './legacy-element-mixin.js';
-import { legacyOptimizations, legacyNoObservedAttributes } from '../utils/settings.js';
-import { findObservedAttributesGetter } from '../mixins/disable-upgrade-mixin.js';
-import { wrap } from '../utils/wrap.js';
-
-const DISABLED_ATTR = 'disable-upgrade';
+import { legacyOptimizations } from '../utils/settings.js';
 
 const lifecycleProps = {
   attached: true,
@@ -180,7 +176,6 @@ function mergeProperties(target, source) {
 }
 
 const LegacyElement = LegacyElementMixin(HTMLElement);
-const observedAttributesGetter = findObservedAttributesGetter(LegacyElement);
 
 /* Note about construction and extension of legacy classes.
   [Changed in Q4 2018 to optimize performance.]
@@ -286,56 +281,12 @@ function GenerateClassFromInfo(info, Base, behaviors) {
      * @return {void}
      */
     created() {
-      // Pull all attribute values 1x if `legacyNoObservedAttributes` is set.
-      if (legacyNoObservedAttributes && this.hasAttributes()) {
-        const a = this.attributes;
-        for (let i=0, l=a.length; i < l; i++) {
-          const attr = a[i];
-          this.__attributeReaction(attr.name, null, attr.value);
-        }
-      }
       super.created();
       const list = lifecycle.created;
       if (list) {
         for (let i=0; i < list.length; i++) {
           list[i].call(this);
         }
-      }
-    }
-
-    /**
-     * Processes an attribute reaction when the `legacyNoObservedAttributes`
-     * setting is in use.
-     * @param {string} name Name of attribute that changed
-     * @param {?string} old Old attribute value
-     * @param {?string} value New attribute value
-     * @return {void}
-     */
-    __attributeReaction(name, old, value) {
-      if ((this.__dataAttributes && this.__dataAttributes[name]) || name === DISABLED_ATTR) {
-        this.attributeChangedCallback(name, old, value);
-      }
-    }
-
-    /** @override */
-    setAttribute(name, value) {
-      if (legacyNoObservedAttributes) {
-        const oldValue = this.getAttribute(name);
-        super.setAttribute(name, value);
-        this.__attributeReaction(name, oldValue, value);
-      } else {
-        super.setAttribute(name, value);
-      }
-    }
-
-    /** @override */
-    removeAttribute(name) {
-      if (legacyNoObservedAttributes) {
-        const oldValue = this.getAttribute(name);
-        super.removeAttribute(name);
-        this.__attributeReaction(name, oldValue, null);
-      } else {
-        super.removeAttribute(name);
       }
     }
 
@@ -473,101 +424,6 @@ function GenerateClassFromInfo(info, Base, behaviors) {
         }
       }
     }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    constructor() {
-      super();
-      /** @type {boolean|undefined} */
-      this.__isUpgradeDisabled;
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    /** @override */
-    static get observedAttributes() {
-      return legacyNoObservedAttributes ? [] :
-        observedAttributesGetter.call(this).concat(DISABLED_ATTR);
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    // Prevent element from initializing properties when it's upgrade disabled.
-    /** @override */
-    _initializeProperties() {
-      if (this.hasAttribute(DISABLED_ATTR)) {
-        this.__isUpgradeDisabled = true;
-      } else {
-        super._initializeProperties();
-      }
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    // Prevent element from enabling properties when it's upgrade disabled.
-    // Normally overriding connectedCallback would be enough, but dom-* elements
-    /** @override */
-    _enableProperties() {
-      if (!this.__isUpgradeDisabled) {
-        super._enableProperties();
-      }
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    // If the element starts upgrade-disabled and a property is set for
-    // which an accessor exists, the default should not be applied.
-    // This additional check is needed because defaults are applied via
-    // `_initializeProperties` which is called after initial properties
-    // have been set when the element starts upgrade-disabled.
-    /** @override */
-    _canApplyPropertyDefault(property) {
-      return super._canApplyPropertyDefault(property) &&
-        !(this.__isUpgradeDisabled && this._isPropertyPending(property));
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    /**
-     * @override
-     * @param {string} name Attribute name.
-     * @param {?string} old The previous value for the attribute.
-     * @param {?string} value The new value for the attribute.
-     * @param {?string} namespace The XML namespace for the attribute.
-     * @return {void}
-     */
-    attributeChangedCallback(name, old, value, namespace) {
-      if (name == DISABLED_ATTR) {
-        // When disable-upgrade is removed, intialize properties and
-        // provoke connectedCallback if the element is already connected.
-        if (this.__isUpgradeDisabled && value == null) {
-          super._initializeProperties();
-          this.__isUpgradeDisabled = false;
-          if (wrap(this).isConnected) {
-            super.connectedCallback();
-          }
-        }
-      } else {
-        super.attributeChangedCallback(
-            name, old, value, /** @type {null|string} */ (namespace));
-      }
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    // Prevent element from connecting when it's upgrade disabled.
-    // This prevents user code in `attached` from being called.
-    /** @override */
-    connectedCallback() {
-      if (!this.__isUpgradeDisabled) {
-        super.connectedCallback();
-      }
-    }
-
-    // NOTE: Inlined for perf from version of DisableUpgradeMixin.
-    // Prevent element from disconnecting when it's upgrade disabled.
-    // This avoids allowing user code `detached` from being called without a
-    // paired call to `attached`.
-    /** @override */
-    disconnectedCallback() {
-      if (!this.__isUpgradeDisabled) {
-        super.disconnectedCallback();
-      }
-    }
-
   }
 
   // apply behaviors, note actual copying is done lazily at first instance creation
